Restructured the tests a little bit to differentiate between Unit and System tests
